move: 데탑/모바일용 레이아웃 컴포넌트 app/layouts 폴더로 분리
